chore: split help command code
